requirements.txt
# pytorch related
torch
torchvision
pytorch-lightning
torchmetrics
efficientnet-pytorch

# data processing
pandas
seaborn
tqdm
ipykernel

# face detection
mediapipe
opencv-python
Pillow

# app
streamlit
streamlit-lottie

# root setup
pyrootutils

# logging
wandb
